Add benchmark steps for evaluating performance fix
This adds fine-grained benchmarking around the rsync operations that
restore repository networks. An additional environment variable
SIMULATE_REPLICAS can be set to reproduce behavior similar to the
performance fix in the previous commit by adding redundant rsync tasks
as if that numbers of replicas had been present at the time that the
backup snapshot was taken.